Improved types part 1

+++ Engine/include/Engine/Entity.h
@@ -8,13 +8,13 @@ namespace engine {
 /// @brief Base class representing any object in the world with a position and size.
 class Entity : public ISettingsObserver {
 public:
-    using Type = int;
+    using Type = uint16_t;
 
     /// @brief Configuration structure used for constructing entities.
     struct Config {
-        utils::f2v pos; ///< Initial position of the entity.
-        utils::f2v size;///< Size of the entity.
-        Type type;      ///< Entity type identifier.
+        utils::f2v pos;  ///< Initial position of the entity.
+        utils::f2v size; ///< Size of the entity.
+        Type type;       ///< Entity type identifier.
     };
 
     Entity(Settings&, const Config& config);
@@ -34,13 +34,13 @@ public:
 protected:
     Type getType() const { return _type; }
 
-    size_t getTileSize() const { return _tileSize; }
+    uint32_t getTileSize() const { return _tileSize; }
 
     const Settings& getSettings() const { return _settings; }
 
 private:
     Settings& _settings; ///< Reference to shared engine settings.
-    size_t _tileSize;    ///< Cached tile size.
+    uint32_t _tileSize;  ///< Cached tile size.
     utils::f2v _pos;     ///< Current position of the entity.
     utils::f2v _size;    ///< Current size of the entity.
     Type _type;          ///< Identifier for the type of entity.
